Factor out driving the LLM.
Also played with the prompt a little. It still needs work. I think I
probebly need to include history so it has examples of what to do. And
maybe add an example of my own that gets pushed out of the history
over time, like the dialog examples in SillyTavern.